Move shared unified encoding code to another class
